output fix

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -11,7 +11,7 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 	int i = 0;
-	unsigned int xor;
+	unsigned long int xor;
 
 	xor = n ^ m;
 	while (xor)
